requirements.txt
streamlit
crewai
groq
litellm
requests
python-dotenv
chromadb
sentence-transformers
bcrypt
cryptography
